+++ src/CosmosDB/CosmosDB/ChangeLog.md
@@ -20,6 +20,7 @@
 
 ## Upcoming Release
 * Upgraded Azure.Core to 1.44.1.
+* Introduced support for creating containers with Vector Embedding Policy.
 
 ## Version 1.15.0
 * Added new parameter `DisableTtl` to `Restore-AzCosmosDBAccount`.